fix: address 10 minor clippy warnings (oag-0dx0)
- Remove unused DoorState import (moved to test module)
- Fix needless borrows: &weapon -> weapon in combat resolution
- Combine consecutive .replace() calls with char array
- Use push('\n') instead of push_str("\n")
- Use is_multiple_of(2) instead of % 2 == 0
- Remove unnecessary u32 cast (combat_hd already returns u32)
- Use io::Error::other() shorthand
- Use is_none_or() instead of map_or(true, ...)
- Fix doc list continuation with blank line

+++ src/engine/combat.rs
@@ -660,7 +660,7 @@ pub fn resolve_turn_undead_with<R: Rng>(
             if !m.is_alive() || m.turned {
                 continue;
             }
-            let m_hd = m.hit_dice.combat_hd().max(1) as u32;
+            let m_hd = m.hit_dice.combat_hd().max(1);
             if m_hd <= remaining_hd {
                 remaining_hd -= m_hd;
                 if destroyed {
